Fix issue with promoted build plugin
The svn-tag plugin takes the environment variables from the ancestor build
and not the actual build that was triggered. The issue can be solved if the
environment variables from the actual build are put in the var map and then
the ancestor build adds/overrides its environment variables to the map. That
would not change the original behaviour of the svn-tag plugin but add the
wanted functionality for it to work well together with the promoted build plugin.

This is a fix for JENKINS-14606

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>